chore: move getFullAppData to cowswap-utils
