+++ packages/swapper/src/cowswap-utils/index.ts
@@ -1,3 +1,6 @@
+import type { LatestAppDataDocVersion } from '@cowprotocol/app-data'
+import { MetadataApi, stringifyDeterministic } from '@cowprotocol/app-data'
+import type { OrderClass, OrderClass1 } from '@cowprotocol/app-data/dist/generatedTypes/v1.3.0'
 import type { ChainId } from '@shapeshiftoss/caip'
 import { ASSET_NAMESPACE, fromChainId, toAssetId } from '@shapeshiftoss/caip'
 import type { EvmChainAdapter, SignTypedDataInput } from '@shapeshiftoss/chain-adapters'
@@ -5,11 +8,16 @@ import { toAddressNList } from '@shapeshiftoss/chain-adapters'
 import type { ETHSignTypedData, HDWallet } from '@shapeshiftoss/hdwallet-core'
 import type { AccountMetadata, TypedDataTypes, UnsignedOrderCreation } from '@shapeshiftoss/types'
 import { CowNetwork, KnownChainIds, TypedDataPrimaryType } from '@shapeshiftoss/types'
-import { bnOrZero, getNativeFeeAssetReference } from '@shapeshiftoss/utils'
+import {
+  bnOrZero,
+  convertDecimalPercentageToBasisPoints,
+  getNativeFeeAssetReference,
+} from '@shapeshiftoss/utils'
 import type { TypedData } from 'eip-712'
 import type { TypedDataDomain } from 'ethers'
 import { ethers } from 'ethers'
 import type { Address } from 'viem'
+import { keccak256, stringToBytes } from 'viem'
 
 import type { AffiliateAppDataFragment } from '../swappers/CowSwapper'
 import { COW_SWAP_SETTLEMENT_ADDRESS } from '../swappers/CowSwapper'
@@ -194,4 +202,47 @@ export const getAffiliateAppDataFragmentByChainId = ({
       recipient: getTreasuryAddressFromChainId(chainId),
     },
   }
+}
+
+type AppDataInfo = {
+  doc: LatestAppDataDocVersion
+  fullAppData: string
+  appDataKeccak256: string
+  env?: string
+}
+
+const generateAppDataFromDoc = async (
+  doc: LatestAppDataDocVersion,
+): Promise<Pick<AppDataInfo, 'fullAppData' | 'appDataKeccak256'>> => {
+  const appData = await stringifyDeterministic(doc)
+  const appDataKeccak256 = keccak256(stringToBytes(appData))
+
+  return { fullAppData: appData, appDataKeccak256 }
+}
+
+const metadataApi = new MetadataApi()
+
+// See https://api.cow.fi/docs/#/default/post_api_v1_quote / https://github.com/cowprotocol/app-data
+export const getFullAppData = async (
+  slippageTolerancePercentage: string,
+  affiliateAppDataFragment: AffiliateAppDataFragment,
+  orderClass1: OrderClass1,
+) => {
+  const APP_CODE = 'shapeshift'
+  const orderClass: OrderClass = { orderClass: orderClass1 }
+  const quote = {
+    slippageBips: convertDecimalPercentageToBasisPoints(slippageTolerancePercentage).toNumber(),
+  }
+
+  const appDataDoc = await metadataApi.generateAppDataDoc({
+    appCode: APP_CODE,
+    metadata: {
+      quote,
+      orderClass,
+      ...affiliateAppDataFragment,
+    },
+  })
+
+  const { fullAppData, appDataKeccak256 } = await generateAppDataFromDoc(appDataDoc)
+  return { appDataHash: appDataKeccak256, appData: fullAppData }
 }
